+++ dbt/adapters/sqlserver/__init__.py
@@ -5,7 +5,7 @@ from dbt.adapters.sqlserver.sql_server_column import SQLServerColumn
 from dbt.adapters.sqlserver.sql_server_configs import SQLServerConfigs
 from dbt.adapters.sqlserver.sql_server_connection_manager import SQLServerConnectionManager
 from dbt.adapters.sqlserver.sql_server_credentials import SQLServerCredentials
-from dbt.include import sqlserver
+from dbt.adapters import sqlserver
 
 Plugin = AdapterPlugin(
     adapter=SQLServerAdapter,
