+++ dbt/adapters/sqlserver/__version__.py
@@ -1,1 +1,1 @@
-version = "1.7.4"
+version = "1.8.0"
